requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
berserk==0.13.2
certifi==2024.8.30
charset-normalizer==3.4.0
chess==1.11.1
chess.com==3.7.0
cloudscraper==1.2.71
darkdetect==0.8.0
Deprecated==1.2.15
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
ndjson==0.3.1
numpy==2.1.3
numpy-stl==3.1.2
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-utils==3.9.0
qrcode==8.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.1
